This commit message was created for the odefa repository.  It was preserved here for historical reasons but may not reflect the rationale or content of this commit. (ZEP - 2016-08-08)
Revert "Work on a closure that only closes over "live" nops -- those which are known to be sourced from a node which has been explicitly marked for queries."
This commit message was created for the odefa repository.  It was preserved here for historical reasons but may not reflect the rationale or content of this commit. (ZEP - 2016-08-08)


This commit message was created for the odefa repository.  It was preserved here for historical reasons but may not reflect the rationale or content of this commit. (ZEP - 2016-08-08)

This reverts commit b471116ad685a9eda1bca60f98a9426e1ddebb63.